adding user registration
