+++ api/index.js
@@ -1,17 +1,7 @@
-const express = require("express");
-const router = express.Router();
-const ctrlContact = require("../controller");
+const contactsRouter = require("./contacts");
+const usersRouter = require("./users");
 
-router.get("/", ctrlContact.getContacts);
-
-router.get("/:contactId", ctrlContact.getContactById);
-
-router.post("/", ctrlContact.createNewContact);
-
-router.delete("/:contactId", ctrlContact.deleteContact);
-
-router.put("/:contactId", ctrlContact.updateContact);
-
-router.patch("/:contactId/favorite", ctrlContact.updateStatusContact);
-
-module.exports = router;
+module.exports = {
+  contactsRouter,
+  usersRouter,
+};
