+++ service/schemas/user.js
@@ -1,4 +1,6 @@
 const mongoose = require("mongoose");
+const bCrypt = require("bcryptjs");
+
 const Schema = mongoose.Schema;
 
 const user = new Schema({
@@ -21,6 +23,14 @@ const user = new Schema({
     default: null,
   },
 });
+
+user.methods.setPassword = function (password) {
+  this.password = bCrypt.hashSync(password, bCrypt.genSaltSync(6));
+};
+
+user.methods.validPassword = function (password) {
+  return bCrypt.compareSync(password, this.password);
+};
 
 const User = mongoose.model("user", user);
